+++ src/Silex/Provider/SecurityServiceProvider.php
@@ -543,7 +543,7 @@ class SecurityServiceProvider implements ServiceProviderInterface
         foreach ($this->fakeRoutes as $route) {
             list($method, $pattern, $name) = $route;
 
-            $app->$method($pattern, null)->bind($name);
+            $app->$method($pattern)->run(null)->bind($name);
         }
     }